Removed cursor-based operations; don't store full Blocks.
I have the feeling that having a mutable Cursor variable on each
Stream structure that keeps track of a single read position would
be a bit of an obstacle for concurrent operations.

Also not prefetching all of the Blob data for a Block, since they
use up a lot of memory.

I ended up opting for fetching everything exactly as it's needed,
each time it's needed, including FileHeaders.

Not the ideal balance, but it's better than the previous strategy.

Optimizing to reduce the number of requests might make a lot of
sense depending on network latency.