HTTP 415 Status Code = io.EOF
415 is the code for "RangeNotSatisfiable," which means that the
bytes requested are not available on the server.

Most likely it's because an additional block is being requested
after the last block in a file.

Though the ideal way to handle this is to cache the file header,
store its total size, and reject requests above that without
even trying to query the server.