Streams: HTTP/S3 now read the entire response Body.
Not the most efficient memory-wise, but there will have to be a
rearch to make it so the consumer of the data doesn't have to
handle closing the Response itself.

Probably at some point a channel that just sends the data chunk by
chunk is the best way to go.